+++ src/main/groovy/ru/vyarus/gradle/plugin/lib/JavaLibPlugin.groovy
@@ -4,12 +4,16 @@ import groovy.transform.CompileStatic
 import groovy.transform.TypeCheckingMode
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.internal.FeaturePreviews
 import org.gradle.api.java.archives.Attributes
 import org.gradle.api.plugins.GroovyPlugin
 import org.gradle.api.plugins.JavaPlugin
 import org.gradle.api.publish.maven.MavenPublication
 import org.gradle.api.tasks.bundling.Jar
+import org.gradle.util.GradleVersion
 import ru.vyarus.gradle.plugin.pom.PomPlugin
+
+import javax.inject.Inject
 
 /**
  * Plugin performs common configuration for java or groovy library:
@@ -35,6 +39,14 @@ class JavaLibPlugin implements Plugin<Project> {
     private static final String JAVADOC_JAR = 'javadocJar'
     private static final String GROOVYDOC_JAR = 'groovydocJar'
 
+    private final FeaturePreviews previews
+
+    // plugin will fail to initialize on gradle before 4.6
+    @Inject
+    JavaLibPlugin(FeaturePreviews featurePreviews) {
+        this.previews = featurePreviews
+    }
+
     @Override
     void apply(Project project) {
         // activated only when java plugin is enabled
@@ -45,7 +57,11 @@ class JavaLibPlugin implements Plugin<Project> {
             addSourcesJarTask(project)
             addJavadocJarTask(project)
             addGroovydocJarTask(project)
-            configureMavenPublication(project)
+            if (isStablePublishingEnabled()) {
+                configureStableMavenPublication(project)
+            } else {
+                configureMavenPublication(project)
+            }
             addInstallTask(project)
         }
     }
@@ -139,8 +155,25 @@ class JavaLibPlugin implements Plugin<Project> {
         }
     }
 
+    private boolean isStablePublishingEnabled() {
+        if (GradleVersion.current() >= GradleVersion.version('5.0')) {
+            // since 5.0 stable publishing should be enabled by default
+            return true
+        }
+        // option available from gradle 4.8
+        try {
+            FeaturePreviews.Feature stablePublishingFeature = FeaturePreviews.Feature
+                    .withName('STABLE_PUBLISHING')
+            return stablePublishingFeature.isActive() && !previews.isFeatureEnabled(stablePublishingFeature)
+        } catch (IllegalArgumentException ignored) {
+            // do nothing if option doesn't exists anymore (.withName() failed)
+            return false
+        }
+    }
+
     private void configureMavenPublication(Project project) {
         project.configure(project) {
+            // lazy initialized publication (could be directly overridden)
             publishing {
                 publications {
                     maven(MavenPublication) {
@@ -151,6 +184,32 @@ class JavaLibPlugin implements Plugin<Project> {
                         }
                         if (project.tasks.findByName(GROOVYDOC_JAR)) {
                             artifact groovydocJar
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    @SuppressWarnings('NestedBlockDepth')
+    private void configureStableMavenPublication(Project project) {
+        project.configure(project) {
+            publishing {
+                publications {
+                    maven(MavenPublication) {
+                        // apply configuration only if it wasn't configured manually,
+                        // for example like publishing.publications.maven.artifacts = [jar, javadocJar]
+                        afterEvaluate {
+                            if (artifacts.empty) {
+                                artifact jar
+                                artifact sourcesJar
+                                if (project.tasks.findByName(JAVADOC_JAR)) {
+                                    artifact javadocJar
+                                }
+                                if (project.tasks.findByName(GROOVYDOC_JAR)) {
+                                    artifact groovydocJar
+                                }
+                            }
                         }
                     }
                 }