+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+* Gradle 4.8 [STABLE_PUBLISHING](https://docs.gradle.org/4.8/userguide/publishing_maven.html#publishing_maven:deferred_configuration) support
+    - Plugin requires gradle 4.6 or above (will fail on earlier gradle).
+    - Gradle 4.6, 4.7 - legacy mode (as before)
+    - (breaking) Gradle 4.8, <5.0 - updated pom plugin (1.3.0) will automatically enable STABLE_PUBLISHING mode         
+    - Gradle 5.0 and above - assumed stable publishing enabled by default
+
+Important: when STABLE_PUBLISHING is enabled (gradle 4.8 and above) publishing configurations will NOT work 
+in a lazy way as before. Use `afterEvaluate {}` INSIDE publication configuration in order to configure lazy properties               
+
 ### 1.0.5 (2017-08-15)
 * Update pom plugin (support gradle java-library plugin)